GameChannel is no longer responsible of monitoring heroes
In case of heroes being restarted there's little to gain forcing the
channel to reconnect to get the new position. During a normal game
with players constantly moving it's ok to wait for the next move to
get the update.

The downside of this approach is that it doesn't work when the player
just fires. Their position would not match with the attacked area.

In any case, this "feature" can be introduced again later when channel
monitoring is completed. Probably transforming the upcoming
ChannelWatcher into a more generalized GameWatcher.